Introduce sge utility
